More UI stubs for testing; added simple roadmap.
Several little bugs fixed as well as I'm getting more code coverage from
the UI.

I think NotifyPoller is too complex to warrant the savings in calls to
the server; it has 4 cases to deal with when registering someone.  And
I've found several subtle bugs in the week or so that it has existed.
I'm going to replace it with a simple one-call-per-entity notification
system.