Vastly simplified notification system.
The only downside is that where we used to poll the server once per
interval no matter how many listeners we had for how many entities, we
now poll for each different entity listened to (once, no matter how many
listeners we have.)

The upside is that the code is much simpler, and we don't run into the
bugs we seem to have with the entity list API not quite getting
changes-since correct.  It also feels less wasteful to only ask the
server about the entities that we are interested in, rather than asking
about all changes and then ignoring most of them.