Server delete supported in new UI
Created a "PaginatedTable" class that sucks a page of entities
at a time into an HTML table.

Still looks ugly, but now it's just due to a lack of CSS styles.

I took a long time trying to figure out how I wanted to display the list
of servers and the details for servers, and decided to punt: for now
I'll just display them as two divs, but the code has been arranged to
allow us to swap between styles fairly simply if we care to experiment.