Server update and create supported in UI.
I feel a bit like I'm reinventing the wheel here, but it may just be
because I've built a few too many CRUD screens in Javascript in my life.

I looked into ExtJS but it felt really heavy for our needs.  I think it
will be best to keep the design really simple for now, and if we can't
jazz up the UI enough with CSS we'll look into using ExtJS or some other
UI framework.